+++ src/__init__.py
@@ -1,8 +1,8 @@
 import sys
 import os
 
-if sys.version_info < (3, 9):
-    raise ImportError("nanobind does not support Python < 3.9.")
+if sys.version_info < (3, 10):
+    raise ImportError("nanobind does not support Python < 3.10.")
 
 def source_dir() -> str:
     "Return the path to the nanobind source directory."
@@ -16,7 +16,7 @@ def cmake_dir() -> str:
     "Return the path to the nanobind CMake module directory."
     return os.path.join(os.path.abspath(os.path.dirname(__file__)), "cmake")
 
-__version__ = "2.16.0-dev1"
+__version__ = "3.0.0-dev1"
 
 __all__ = (
     "__version__",