+++ src/nb_type.cpp
@@ -895,7 +895,7 @@ static constexpr nb_slot type_slots[] {
     E(78, as_async, am, aiter),
     E(79, as_async, am, anext),
     E(80, ht_type, tp, finalize),
-#if PY_VERSION_HEX >= 0x030A0000 && !defined(PYPY_VERSION)
+#if !defined(PYPY_VERSION)
     E(81, as_async, am, send),
 #endif
 };
@@ -1208,7 +1208,7 @@ PyTypeObject *nb_type_create_metaclass(nb_internals *p,
         /* .name = */ "nanobind.nb_type",
         /* .basicsize = */ basicsize,
         /* .itemsize = */ itemsize,
-        /* .flags = */ Py_TPFLAGS_DEFAULT | NB_TPFLAGS_IMMUTABLETYPE,
+        /* .flags = */ Py_TPFLAGS_DEFAULT | Py_TPFLAGS_IMMUTABLETYPE,
         /* .slots = */ slots
     };